Remove team user team creation for sso users
